fix: added middlewares folder for protected routes
